Changes instances_from_po_lines for new data from bookplate_funds_polines and moves tests to renamed test_bookplates.py.
